requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.2.3
gunicorn==21.2.0
gspread==6.1.2
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
